remove another cast.

+++ src/types.h
@@ -273,7 +273,7 @@ inline Value eg_value(Score s) {
 }
 
 inline Value mg_value(Score s) {
-  return Value((int16_t)(uint16_t)(unsigned)(s & 0xFFFF));
+  return Value((int16_t)(unsigned)(s & 0xFFFF));
 }
 
 #define ENABLE_BASE_OPERATORS_ON(T)                                \
